refactor : LandingMember 컴포넌트 분리
1. 유저 상태를 표시하는 블록 UserBlock 분리
2. LandingMember를 별도의 컴포넌트로 분리
